Add startInputOutput(), simultaneous DMX input and Output
This update allows you to receive DMX data (as a client) on one
universe, while simultaneously sending data (as a controller) on another
universe. This is different from RDM, as this is two unidirectional DMX
interfaces, whereas RDM is one bidirectional interface. This does not
support RDM, but it may be possible to integrate RDM Client funtionality
on the RX (UART0) side.

See DMXInputOutputTest.ino for an example.

Internal changes:
- _receivedData renamed to _receiveBuffer
- when dmx packet receive completes, copies contents of _receiveBuffer
to _receivedDmxData
- replaced _slots with _tx_max_slots and _num_received_slots

API changes:
- add LX8266DMX.startInputOutput() function.
- receivedData() renamed to receiveBuffer(). As before, it returns a
pointer to the buffer that the rx interrupt writes into.
- receivedDMXData() is new. It returns a pointer to a complete received
DMX packet.
- printReceivedData() renamed to printReceiveBuffer()
- getSlot() reads from the _receivedDMXData, which is now different from
_dmxData.

The autogenerated HTML documentation in "extras" folder has not been
updated to reflect the code changes.