Fix intermittent corrupted RX buffer issue during InputOutput
When using startInputOutput, there were intermittent flashes as a result
of RX data buffer corruption. It seems to be a result of handling both
RX and TX interrupt events when both are available. If we instead ignore
the TX interrupt (presumably it fires the interrupt handler again
immediately after exiting) when there is an RX interrupt to handle, the
problem goes away.